Change rule count types to int64_t. SNTRuleIdentifiers properties now RO.

+++ Source/common/SNTXPCUnprivilegedControlInterface.h
@@ -23,12 +23,12 @@
 @class MOLXPCConnection;
 
 struct RuleCounts {
-  NSUInteger binary;
-  NSUInteger certificate;
-  NSUInteger compiler;
-  NSUInteger transitive;
-  NSUInteger teamID;
-  NSUInteger signingID;
+  int64_t binary;
+  int64_t certificate;
+  int64_t compiler;
+  int64_t transitive;
+  int64_t teamID;
+  int64_t signingID;
 };
 
 ///
@@ -46,7 +46,7 @@ struct RuleCounts {
 ///  Database ops
 ///
 - (void)databaseRuleCounts:(void (^)(struct RuleCounts ruleCounts))reply;
-- (void)databaseEventCount:(void (^)(NSUInteger count))reply;
+- (void)databaseEventCount:(void (^)(int64_t count))reply;
 - (void)staticRuleCount:(void (^)(int64_t count))reply;
 
 ///
